+++ expyriment/io/_keyboard.py
@@ -220,6 +220,8 @@ class Keyboard(Input):
 
         """
 
+        if expyriment.control.defaults._skip_wait_functions:
+            return None, None
         if android_show_keyboard is not None:
             android_show_keyboard()
         start = get_time()
@@ -286,6 +288,8 @@ class Keyboard(Input):
 
         """
 
+        if expyriment.control.defaults._skip_wait_functions:
+            return None, None
         start = get_time()
         rt = None
         found_char = None