+++ expyriment/misc/data_preprocessing.py
@@ -18,6 +18,7 @@ except ImportError:
     _locale = None  # Does not exist on Android
 import sys as _sys
 import types as _types
+from tempfile import mkstemp as _mkstemp
 from copy import copy as _copy
 import codecs as _codecs
 import re as _re
@@ -155,7 +156,7 @@ def write_csv_file(filename, data, varnames=None, delimiter=','):
 
 
 def write_concatenated_data(data_folder, file_name, output_file=None,
-                            delimiter=','):
+                            delimiter=',', to_R_data_frame=False):
     """Concatenate data and write it to a csv file.
 
     All files that start with this name will be considered for the
@@ -178,10 +179,17 @@ def write_concatenated_data(data_folder, file_name, output_file=None,
         to {file_name}.csv
     delimiter : str, optional
         delimiter character (default=",")
+    to_R_data_frame: bool, optional
+        if True, data will be converted to a R data frame that is saved
+        in a RDS file
 
     """
 
-    return Aggregator(data_folder=data_folder, file_name=file_name)\
+    if to_R_data_frame:
+        return Aggregator(data_folder=data_folder, file_name=file_name)\
+        .write_concatenated_data_to_R_data_frame(output_file=output_file)
+    else:
+        return Aggregator(data_folder=data_folder, file_name=file_name)\
         .write_concatenated_data(output_file=output_file, delimiter=delimiter)
 
 
@@ -960,13 +968,13 @@ The Python package 'numpy' is not installed."""
         self._last_data = []
 
     def write_concatenated_data(self, output_file=None, delimiter=','):
-        """Concatenate data and write it to a csv file.
+        """Concatenates data and writes it to a csv file.
 
         Parameters
         ----------
         output_file : str, optional
             name of data output file
-            If no specified data will the save to {file_name}.csv
+            If not specified data will the save to {file_name}.csv
         delimiter : str
             delimiter character (default=",")
 
@@ -978,6 +986,45 @@ The Python package 'numpy' is not installed."""
         data = self.concatenated_data
         write_csv_file(filename=output_file, data=data[0], varnames=data[1],
                        delimiter=delimiter)
+
+    def write_concatenated_data_to_R_data_frame(self, output_file):
+        """Creates a R data frame of the concatenated data and stores it in a
+        RDS file.
+
+        Parameters
+        ----------
+        output_file : str, optional
+            name of RDS output file
+            If not specified data will the save to {file_name}.csv
+
+        Notes
+        -----
+        This method requires R and the Python package 'Rpy2'.
+
+        """
+
+        try:
+            import rpy2.robjects as robjects
+        except:
+            message = "Saving data to R data frame requires the " +\
+                    "Python package 'Rpy2', which is not installed."""
+            raise ImportError(message)
+
+        if output_file is None:
+            output_file = u"{0}.rds".format(self.file_name)
+        fl, tmp_file_name = _mkstemp()
+        _os.close(fl)
+        self.write_concatenated_data(output_file=tmp_file_name, delimiter=',')
+
+        robjects.r('''data = read.csv("{0}", comment.char="#",
+        na.strings=c("NA", "None"))'''.format(tmp_file_name))
+        robjects.r('''str(data)''')
+        print("write file: {0}".format(output_file))
+        robjects.r('''saveRDS(data, file="{0}")'''.format(output_file))
+        try:
+            _os.remove(tmp_file_name)
+        except:
+            pass
 
     def set_independent_variables(self, variables):
         """Set the independent variables.